+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### New Features
 
+- Extend binarize target scorers to support stacking estimators, i.e., use binarize target estimator as meta estimator.
 - Adds `cv_results` attributes to `estimator_attributes` tool.